Move `util-core` into the `@bayou` scope.
