+++ local-modules/@bayou/util-core/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@bayou/util-core",
+  "version": "1.0.0"
+}
