+++ local-modules/doc-macros/package.json
@@ -6,6 +6,6 @@
     "deps-quill-client": "local",
     "quill-util": "local",
     "ui-embeds": "local",
-    "util-core": "local"
+    "util-common": "local"
   }
 }
