+++ local-modules/quill-util/NotReallyMarkdown.js
@@ -3,7 +3,7 @@
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
 import { Logger } from 'see-all';
-import { UtilityClass } from 'util-core';
+import { UtilityClass } from 'util-common';
 
 import QuillEvents from './QuillEvents';
