+++ local-modules/typecheck/package.json
@@ -3,6 +3,6 @@
   "version": "1.0.0",
 
   "dependencies": {
-    "util-core": "local"
+    "@bayou/util-core": "local"
   }
 }
