+++ local-modules/ui-embeds/ComponentBlotWrapper.js
@@ -10,7 +10,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { ClientStore } from 'data-model-client';
 import { TBoolean, TFunction, TNumber, TObject, TString } from 'typecheck';
-import { Errors, UtilityClass } from 'util-core';
+import { Errors, UtilityClass } from '@bayou/util-core';
 
 const BlockEmbed = Quill.import('blots/block/embed');
