+++ local-modules/util-common/Singleton.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { CommonBase, Errors, ObjectUtil } from 'util-core';
+import { CommonBase, Errors, ObjectUtil } from '@bayou/util-core';
 
 /**
  * Base class for singletons (classes for which there is ever but a single
