--- local-modules/util-core/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "util-core",
-  "version": "1.0.0"
-}
